refactor: group together helper functions in separate dir
