+++ src/mappings/poolFactory.ts
@@ -1,5 +1,6 @@
-import { ZERO_BD, VAULT_ADDRESS } from './constants';
-import { newPoolEntity, createPoolTokenEntity, scaleDown, loadPoolToken } from './helpers';
+import { ZERO_BD, VAULT_ADDRESS } from './helpers/constants';
+import { newPoolEntity, createPoolTokenEntity, scaleDown, loadPoolToken } from './helpers/misc';
+import { updatePoolWeights } from './helpers/weighted';
 
 import { BigInt, Address, Bytes } from '@graphprotocol/graph-ts';
 import { PoolCreated } from '../types/WeightedPoolFactory/WeightedPoolFactory';
@@ -14,33 +15,6 @@ import { Vault } from '../types/Vault/Vault';
 import { WeightedPool } from '../types/templates/WeightedPool/WeightedPool';
 import { StablePool } from '../types/templates/StablePool/StablePool';
 import { ConvergentCurvePool } from '../types/templates/ConvergentCurvePool/ConvergentCurvePool';
-
-export function updatePoolWeights(poolId: string): void {
-  let pool = Pool.load(poolId);
-  let poolContract = WeightedPool.bind(pool.address as Address);
-
-  let tokensList = pool.tokensList;
-  let weightsCall = poolContract.try_getNormalizedWeights();
-  if (!weightsCall.reverted) {
-    let weights = weightsCall.value;
-    let totalWeight = ZERO_BD;
-
-    for (let i: i32 = 0; i < tokensList.length; i++) {
-      let tokenAddress = tokensList[i] as Address;
-      let weight = weights[i];
-
-      let poolToken = loadPoolToken(poolId, tokenAddress);
-      poolToken.weight = scaleDown(weight, 18);
-      poolToken.save();
-
-      totalWeight = totalWeight.plus(scaleDown(weight, 18));
-    }
-
-    pool.totalWeight = totalWeight;
-  }
-
-  pool.save();
-}
 
 export function handleNewWeightedPool(event: PoolCreated): void {
   let poolAddress: Address = event.params.pool;
